Fix test

+++ src/main/java/at/rtr/rmbt/repository/impl/RadioSignalRepositoryImpl.java
@@ -160,7 +160,7 @@ public class RadioSignalRepositoryImpl implements RadioSignalRepository {
                                 lteRsrp,
                                 lteRsrq,
                                 rsSignal.getString("cat_technology"),
-                                rsSignal.getObject("location_id", Integer.class),
+                                rsSignal.getObject("location_id", Long.class).intValue(),
                                 rsSignal.getObject("area_code", Integer.class),
                                 rsSignal.getObject("primary_scrambling_code", Integer.class),
                                 rsSignal.getObject("channel_number", Integer.class),
